chore: clippy auto-fixes + npm audit fix (frontend 0 vulnerabilities)
Post-merge polish. `cargo clippy --fix` applied mechanical, semantically-preserving
lint fixes (uninlined format args, redundant closures, etc.) across 13 Rust files.
`npm audit fix` resolved the 2 high-severity advisories that arrived with the
dependency modernization on main (React 19 family), with no breaking changes.

Verified: clippy -D correctness -D suspicious (0), workspace unit tests pass,
frontend vitest 649 pass, frontend build (0), npm audit → 0 vulnerabilities.

+++ crates/weissman-agent/src/detections/baseline.rs
@@ -117,7 +117,7 @@ fn collect_metrics() -> Value {
 
     // ── Uptime + load + memory (Linux /proc/uptime, /proc/loadavg, /proc/meminfo)
     if let Some(uptime) = read_uptime_seconds() {
-        m.insert("uptime_seconds".into(), Value::from(uptime as u64));
+        m.insert("uptime_seconds".into(), Value::from(uptime));
     }
     if let Some(load) = read_loadavg_1m() {
         m.insert("load_1m".into(), json!(load));
